requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pgvector
python-dotenv

google-generativeai
openai
httpx
pydantic

